Move convertAutomergeToSlateOps
