+++ src/utils/convertAutomergeToSlateOps.js
@@ -1,0 +1,183 @@
+export const convertAutomergeToSlateOps = (automergeOps, pathMap, value) => {
+  // To build objects from Automerge operations
+  const slateOps = []
+  const objIdMap = {}
+  const deferredOps = []
+
+  automergeOps.map(op => {
+    if (op.action === 'create') {
+      switch (op.type) {
+        case 'map':
+          objIdMap[op.obj] = {}
+          break;
+        case 'list':
+          objIdMap[op.obj] = []
+          break;
+        default:
+          console.error('`create`, unsupported type: ', op.type)
+      }
+    }
+
+    if (op.action === 'remove') {
+      let pathString, slatePath, slateOp
+
+      pathString = pathMap[op.obj].match(/\d+/g)
+      if (pathString) {
+        slatePath = pathString.map(x => {
+          return parseInt(x, 10);
+        });
+      }
+      else {
+        // FIXME: Is `op.index` always the right path? What happens in a
+        // sub-node (in other words, will `slatePath` ever need to have
+        // length > 1?
+        slatePath = [op.index]
+      }
+
+      // Validate the operation using the node type at path given by `op.obj`
+      // FIXME: Is the Slate's Value reliable enough to get the node type?
+      // Use the document to be changed
+      const removeNode = value.document.getNodeAtPath(slatePath)
+      switch (removeNode.object) {
+        case 'text':
+          slateOp = {
+            type: 'remove_text',
+            path: slatePath,
+            offset: op.index,
+            text: '*',
+            marks: []
+          }
+          break;
+        case 'block':
+          slateOp = {
+            type: 'remove_node',
+            path: slatePath,
+            node: removeNode
+          }
+          break;
+        default:
+          console.error('`remove`, unsupported node type: ', removeNode.object)
+      }
+
+      slateOps.push(slateOp)
+    }
+
+    if (op.action === "set") {
+      if (op.hasOwnProperty('link')) {
+        // What's the point of the `link` field? All my experiments
+        // have `link` = true
+        if (op.link) {
+          // Check if linking to a newly created object or one that
+          // already exists in our Automerge document
+          if (objIdMap.hasOwnProperty(op.value)) {
+            objIdMap[op.obj][op.key] = objIdMap[op.value]
+          }
+          else if (pathMap.hasOwnProperty(op.value)) {
+            objIdMap[op.obj][op.key] = pathMap[op.value]
+          }
+          else {
+            // TODO: Does this ever happen?
+            console.error('`set`, unable to find objectId: ', op.value)
+          }
+        }
+      }
+      else {
+        objIdMap[op.obj][op.key] = op.value
+      }
+    }
+
+    if (op.action === 'insert') {
+      if (op.link) {
+        // Check if inserting into a newly created object or one that
+        // already exists in our Automerge document
+        if (objIdMap.hasOwnProperty(op.obj)) {
+          objIdMap[op.obj][op.index] = objIdMap[op.value]
+        }
+        else if (pathMap.hasOwnProperty(op.obj)) {
+          deferredOps.push(op)
+        }
+        else {
+          // TODO: Does this ever happen?
+          console.error('`insert`, unable to find objectId: ', op.obj)
+        }
+      }
+      else {
+        // TODO: Does this ever happen?
+        console.log('op.action is `insert`, but link is false')
+      }
+    }
+  })
+
+  // We know all ops in this list have the following conditions true:
+  //  - op.action === `insert`
+  //  - pathMap.hasOwnProperty(op.obj)
+  //  - typeof pathMap[op.obj] === 'string' ||
+  //    pathMap[op.obj] instanceof String
+  deferredOps.map(op => {
+    const insertInto = pathMap[op.obj]
+
+    let pathString, slatePath
+    let slateOp
+
+    // If the `pathString` is available, then we are likely inserting text
+    // FIXME: Verify this
+    pathString = insertInto.match(/\d+/g)
+    if (pathString) {
+      slatePath = pathString.map(x => {
+        return parseInt(x, 10);
+      });
+
+      slateOp = {
+        type: 'insert_text',
+        path: slatePath,
+        offset: op.index,
+        text: objIdMap[op.value].text,
+        marks: objIdMap[op.value].marks
+      }
+    }
+    else {
+      // FIXME: Is `op.index` always the right path? What happens in a
+      // sub-node?
+      slatePath = [op.index]
+
+      // 5/27/18: `insert_node` can't seem to insert a node with pre-existing
+      // text, so we need to insert a node, then `insert_text` into that node
+
+      // Extract text from node to insert, then insert a "clean node", and
+      // re-insert text with `insert_text`
+      const insertNode = objIdMap[op.value]
+      const insertTextNodes = insertNode.nodes
+
+      insertNode.nodes = [{
+        object: 'text',
+        characters: []
+      }]
+
+      slateOp = {
+        type: 'insert_node',
+        path: slatePath,
+        node: insertNode
+      }
+      slateOps.push(slateOp)
+
+      // TODO: Convert the `Text` object properly into separate `insert_text`
+      // operations with proper marks
+      const nodeTextString = insertTextNodes.map(textNode => {
+        return textNode.characters.map(character => {
+          return character.text
+        }).join('')
+      })
+      slateOp = {
+        type: 'insert_text',
+        // Insert the text in the first node of the newly created node
+        path: [slatePath[0], 0],
+        offset: 0,
+        text: nodeTextString.join(''),
+        marks: []
+      }
+    }
+
+    slateOps.push(slateOp)
+  })
+  return slateOps;
+}
